added test for hexstr
find out unity test only accept one test file so renamed test file
